+++ tests/fixtures/layout/relative-context.vue
@@ -1,0 +1,15 @@
+<template>
+  <article class="test-relative-context">
+    <span class="icon">!</span>
+  </article>
+</template>
+<style scoped>
+.test-relative-context {
+  position: relative;
+  > .icon {
+    position: absolute;
+    inset-block-start: 0;
+    inset-inline-end: 0;
+  }
+}
+</style>